update for ubuntu

+++ symlink.sh
@@ -1,20 +1,19 @@
 #!/usr/bin/env bash
+[[ -s ~/.vimrc ]] && mv ~/.vimrc ~/.vimrc.original
+[[ -s ~/.gvimrc ]] && mv ~/.gvimrc ~/.gvimrc
+[[ -s ~/.bashrc ]] && mv ~/.bashrc  ~/.bashrc.original
+[[ -s ~/.bash_profile ]] && mv ~/.bash_profile  ~/.bash_profile.original
+[[ -s ~/.myprofile ]] && mv ~/.myprofile ~/.myprofile.original
+[[ -s ~/.tmux.conf ]] && mv ~/.tmux.conf ~/.tmux.conf.original
+[[ -s ~/.gitconfig ]] && mv ~/.gitconfig  ~/.gitconfig.original
+[[ -s ~/.gitignore ]] && mv ~/.gitignore ~/.gitignore.original
+[[ -s ~/.eslintrc ]] && mv ~/.eslintrc  ~/.eslintrc.original
+[[ -s ~/.aliases ]] && mv ~/.aliases ~/.aliases.original
+[[ -s ~/.zshrc ]] && mv ~/.zshrc ~/.zshrc.original
 
-rm ~/.vimrc
-rm ~/.gvimrc
-rm ~/.zshrc
-rm ~/.bashrc
-rm ~/.bash_profile
-rm ~/.myprofile
-rm ~/.gitconfig
-rm ~/.gitignore
-rm ~/.tmux.conf
-rm ~/.eslintrc
-rm ~/.aliases
-mv ~/.zpreztorc ~/.zpreztorc.original
-mv ~/.zshrc ~/.zshrc.original
+mkdir -p ~/.config/nvim/
 
-cwd=${PWD} 
+cwd=${PWD}
 ln -fs $cwd/.vimrc ~/.vimrc
 ln -fs $cwd/.gvimrc ~/.gvimrc
 ln -fs $cwd/zprezto/.zpreztorc ~/.zpreztorc
@@ -27,3 +26,4 @@ ln -fs $cwd/.gitignore ~/.gitignore
 ln -fs $cwd/.tmux.conf ~/.tmux.conf
 ln -fs $cwd/.eslintrc ~/.eslintrc
 ln -fs $cwd/.aliases ~/.aliases
+ln -fs $cwd/nvim/init.vim ~/.config/nvim/
